requirements.txt
python-dotenv
requests
pymongo
nltk
pandas
textblob
sentence_transformers
datetime
importlib

